requirements.txt
apt-xapian-index==0.49
asgiref==3.2.7
attrs==19.3.0
beautifulsoup4==4.8.2
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
command-not-found==0.3
cryptography==2.9.2
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.0.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
entrypoints==0.3
future==0.18.2
GDAL==3.0.4
gpg===1.13.1-unknown
html5lib==1.0.1
httplib2==0.14.0
idna==2.9
importlib-metadata==1.5.0
ipython-genutils==0.2.0
Jinja2==2.10.1
jsonschema==3.2.0
jupyter-core==4.6.3
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.5.0
MarkupSafe==1.1.0
matplotlib==3.1.2
more-itertools==4.2.0
nbformat==5.0.4
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
OWSLib==0.19.1
pexpect==4.6.0
Pillow==7.0.0
plotly==4.4.1
psycopg2==2.8.4
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyproj==2.5.0
PyQt5==5.14.1
pyrsistent==0.15.5
python-apt==2.0.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.34
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
retrying==1.3.3
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
soupsieve==1.9.5
sqlparse==0.3.1
systemd-python==234
traitlets==4.3.3
typing==3.7.4.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.7
wadllib==1.3.3
webencodings==0.5.1
xkit==0.0.0
zipp==1.0.0
